Convert filter components to TypeScript
