+++ packages/ndla-ui/src/Filter/filterClasses.ts
