+++ packages/ndla-ui/src/Frontpage/FrontpageAllSubjects.tsx
@@ -5,7 +5,6 @@ import Tabs from '@ndla/tabs';
 import SafeLink from '@ndla/safelink';
 import { colors, fonts, mq, breakpoints } from '@ndla/core';
 import { MessageBox } from '../Messages';
-// @ts-ignore
 import { ToggleItem } from '../Filter';
 
 const StyledWrapper = styled.nav`
