+++ packages/ndla-ui/src/Subject/index.ts
@@ -4,7 +4,6 @@ export { default as SubjectLinks } from './SubjectLinks';
 export { default as SubjectArchive } from './SubjectArchive';
 export { default as SubjectAbout } from './SubjectAbout';
 export { default as SubjectCarousel } from './SubjectCarousel';
-export { default as SubjectFilter } from './SubjectFilter';
 export { default as SubjectNewContent } from './SubjectNewContent';
 export { default as SubjectBanner } from './SubjectBanner';
