+++ packages/ndla-ui/src/index-javascript.js
@@ -6,8 +6,6 @@
  *
  */
 
-export { FilterList, FilterListPhone, FilterButtons } from './Filter';
-
 export { default as TopicIntroductionList } from './TopicIntroductionList';
 export { default as TopicMenu, TopicMenuButton } from './TopicMenu';
 export {
