+++ packages/ndla-ui/src/index.ts
@@ -81,6 +81,8 @@ export {
   NdlaFilmHero,
 } from './Hero';
 
+export { FilterList, FilterListPhone, FilterButtons } from './Filter';
+
 export { Footer, EditorName, FooterText, FooterAuth } from './Footer';
 
 export {
@@ -208,7 +210,6 @@ export {
   SubjectCarousel,
   SubjectChildContent,
   SubjectContent,
-  SubjectFilter,
   SubjectFlexChild,
   SubjectFlexWrapper,
   SubjectHeader,
